# assets/openapi/sonos-hub.v1.yaml
openapi: 3.0.3
info:
  title: Sonos Smart Routines Hub API
  version: '1.0.0'
  description: |
    REST API for the Sonos Smart Routines Hub - a LAN-only home automation system for
    Sonos speakers. Provides device discovery, scene/routine management, music playback
    control, and integration with Apple Music. Designed for local-first operation with
    observability via Prometheus metrics and structured JSON logging.
servers:
  - url: http://hub.local:8787
...
paths:
...
  /v1/devices/{device_id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      description: Retrieve detailed information about a specific device
      parameters:
        - in: path
          name: device_id
          description: Unique device identifier
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceResponse' }
        '404':
          description: Device not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # =========================================================================
  # SONOS ENDPOINTS
  # =========================================================================
...
components:
  schemas:
...
    Remediation:
      type: object
      description: Remediation guidance for error recovery
      required: [action]
      properties:
        action:
          type: string
          description: Machine-readable action code (e.g., 'BOOTSTRAP_SERVICE', 'RETRY_LATER')
        endpoint:
          type: string
          description: API endpoint to call to fix the issue
        user_action:
          type: string
          description: Human-readable instruction for the user
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            remediation: { $ref: '#/components/schemas/Remediation' }
...
    Device:
      type: object
      required:
        [
          device_id,
          room_name,
          ip,
          role,
          is_targetable,
          supports_airplay,
          last_seen_at,
          model,
          is_coordinator_capable,
          logical_group_id,
          physical_device_count
        ]
      properties:
        device_id: { type: string }
        udn: { type: string, nullable: true }
        room_name: { type: string }
        ip: { type: string }
        model: { type: string }
        role: { type: string }
        is_targetable: { type: boolean }
        is_coordinator_capable: { type: boolean }
        supports_airplay: { type: boolean }
        logical_group_id: { type: string, nullable: true }
        last_seen_at: { type: string, format: date-time }
        physical_device_count: { type: integer }
        physical_devices:
          type: array
          items: { $ref: '#/components/schemas/PhysicalDevice' }
    PhysicalDevice:
      type: object
      required:
        [
          device_id,
          udn,
          model,
          model_number,
          room_name,
          role,
          is_coordinator_capable,
          supports_airplay,
          last_seen_at
        ]
      properties:
        device_id: { type: string }
        udn: { type: string }
        model: { type: string }
        model_number: { type: string }
        room_name: { type: string }
        role: { type: string }
        is_coordinator_capable: { type: boolean }
        supports_airplay: { type: boolean }
        last_seen_at: { type: string, format: date-time }
...
    DeviceResponse:
      type: object
      required: [request_id, device]
      properties:
        request_id: { type: string }
        device: { $ref: '#/components/schemas/Device' }
...